requirements.txt
aiohttp
asyncio
discord-py-interactions~=3.0.2
discord.py~=1.7.3
yadisk~=1.2.14
PyYAML~=5.4.1
requests~=2.26.0